requirements.txt
torch
torchvision
lpips
numpy
pillow
tqdm